% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_data_t1s}
\alias{crit_from_data_t1s}
\title{crit_from_data_t1s}
\usage{
crit_from_data_t1s(m, s, n, se = NULL, df = NULL, conf.level, hypothesis)
}
\arguments{
\item{m}{mean of the sample.}

\item{s}{a number corresponding to the variance.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{df}{degrees of freedom.}

\item{conf.level}{confidence level of the interval.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}
}
\value{
the output returns a "d" which is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the cohen's d and the critical d for a one samples t-test given the mean, the variance and the sample numerosity, specifying the standard error, degrees of freedom, confidence level of the interval and the direction of the hypothesis.
}
