% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_data_t2s}
\alias{crit_from_data_t2s}
\title{crit_from_data_t2s}
\usage{
crit_from_data_t2s(
  m1,
  m2,
  sd1,
  sd2,
  n1,
  n2,
  conf.level,
  hypothesis,
  se = NULL,
  df = NULL,
  var.equal = FALSE
)
}
\arguments{
\item{m1}{a number representing the mean of group 1.}

\item{m2}{a number representing the mean of group 2.}

\item{sd1}{a number representing the standard deviation of group 1.}

\item{sd2}{a number representing the standard deviation of group 2.}

\item{n1}{a number corresponding to the sample size of group 1.}

\item{n2}{a number corresponding to the sample size of group 2.}

\item{conf.level}{confidence level of the interval.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{se}{a number corresponding to the standard error.}

\item{df}{degrees of freedom.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.}
}
\value{
the output returns a "d" which is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the cohen's d and the critical d given the mean of the two groups, the standard deviation of the means and sample size of the two groups, specifying the confidence level of the interval, the direction of the hypothesis, the standard error, degrees of freedom and the variance parameter.
}
