% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_data_t2sp}
\alias{crit_from_data_t2sp}
\title{crit_from_data_t2sp}
\usage{
crit_from_data_t2sp(
  m1,
  m2 = NULL,
  sd1,
  sd2 = NULL,
  r12 = NULL,
  n,
  df = NULL,
  conf.level,
  hypothesis
)
}
\arguments{
\item{m1}{a number representing the mean of group 1.}

\item{m2}{a number representing the mean of group 2.}

\item{sd1}{a number representing the standard deviation of group 1.}

\item{sd2}{a number representing the standard deviation of group 2.}

\item{r12}{a number corresponding to the correlation between variable 1 and variable 2.}

\item{n}{a number corresponding to the sample size.}

\item{df}{degrees of freedom.}

\item{conf.level}{confidence level of the interval.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}
}
\value{
the output returns a "dz" which is the d standartized on the standard deviation of the differencens, the "dzc" is the critical standardized d, the "d" is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the standardized cohen's d, the critical standardized cohen's d, cohen's d, critical cohen's d and the numerator of formula from which to compute the cohen's d  given the mean of the two groups, the standard deviation of the means, correlation between the two variables and sample size, specifying the degrees of freedom, the confidence level of the interval and the direction of the hypothesis.
}
