% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_t_t1s}
\alias{crit_from_t_t1s}
\title{crit_from_t_t1s}
\usage{
crit_from_t_t1s(t = NULL, n, se = NULL, conf.level, hypothesis)
}
\arguments{
\item{t}{the t value.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{conf.level}{confidence level of the interval.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}
}
\value{
the output returns a "d" which is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the cohen's d and the critical d for a one samples t-test given the t-value, sample size, specifying the confidence level of the interval and the direction of the hypothesis.
}
