% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{crit_from_t_t2sp}
\alias{crit_from_t_t2sp}
\title{crit_from_t_t2sp}
\usage{
crit_from_t_t2sp(t = NULL, n, se = NULL, r12 = NULL, hypothesis, conf.level)
}
\arguments{
\item{t}{the t value.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{r12}{a number corresponding to the correlation between variable 1 and variable 2.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{confidence level of the interval.}
}
\value{
the output returns a "dz" which is the d standartized on the standard deviation of the differences, the "dzc" is the critical standardized d, the "d" is the cohen's d, the critical d which is the minimum value for which to get a significant result with a given sample, the "bc" is the numerator of the formula from which the d is calculated and "df" are the degrees of freedom.
}
\description{
This function allows to calculate the standardized cohen's d, the critical standardized cohen's d, cohen's d and critical cohen's d given for a paired t test, given the t-value, the sample size, the standard error and the correlation between the two variables, specifying the confidence level of the interval and the direction of the hypothesis.
}
