% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-methods.R
\name{critical.htest}
\alias{critical.htest}
\title{critical.htest}
\usage{
\method{critical}{htest}(x, type = NULL)
}
\arguments{
\item{x}{an object of class htest}
}
\description{
una bellissima funzione
}
