% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-methods.R
\name{critical.lm}
\alias{critical.lm}
\title{Compute critical effect size values for linear regression models}
\usage{
\method{critical}{lm}(x, conf.level = 0.95, test = "t", ...)
}
\arguments{
\item{x}{an object of class lm}

\item{conf.level}{the confidence interval level, needed to compute the smallest significant coefficient (default is 0.95, equaling a critical alpha = 0.05)}

\item{test}{which test should be used. \code{t} for a standard t-test (the default) and \code{z} for a z test.}

\item{...}{Additional arguments (currently unused).}
}
\value{
an object of class critvalue
}
\description{
Compute critical effect size values for linear model coefficients of objects of class lm
}
