% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-methods.R
\name{critical}
\alias{critical}
\title{critical}
\usage{
critical(x, ...)
}
\arguments{
\item{x}{an object of class htest, rma or lm}

\item{...}{other arguments passed for the specific method}
}
\value{
a object of class critvalue
}
\description{
describe the function
}
