% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-methods.R
\name{critical.rma}
\alias{critical.rma}
\title{Compute critical effect size values for meta-analytic linear models}
\usage{
\method{critical}{rma}(x, conf.level = 0.95)
}
\arguments{
\item{x}{an object of class rma}

\item{conf.level}{the confidence interval level, needed to compute the smallest significant coefficient (default is 0.95, equaling a critical alpha = 0.05)}
}
\value{
an object of class critvalue
}
\description{
Compute critical effect size values for coefficients of objects of class rma (fitted with the "metafor" package)
}
