% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{critical_t2s}
\alias{critical_t2s}
\title{critical_t2s}
\usage{
critical_t2s(
  m1 = NULL,
  m2 = NULL,
  t = NULL,
  sd1 = NULL,
  sd2 = NULL,
  n1,
  n2,
  se = NULL,
  df = NULL,
  var.equal = FALSE,
  hypothesis = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{m1}{a number representing the mean of group 1.}

\item{m2}{a number representing the mean of group 2.}

\item{t}{the t value.}

\item{sd1}{a number representing the standard deviation of group 1.}

\item{sd2}{a number representing the standard deviation of group 2.}

\item{n1}{a number corresponding to the sample size of group 1.}

\item{n2}{a number corresponding to the sample size of group 2.}

\item{se}{a number corresponding to the standard error.}

\item{df}{degrees of freedom.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{the confidence level to set the confidence interval, default is set to 0.95.}
}
\value{
the output returns, based on the inserted values, means of the two groups, standard deviations, numerosity of the two groups, confidence interval, hypothesis, standard error, degrees of freedom and the set parameter for var.equal or the t-value, numerosity of the two groups, standard error, confidence interval, hypothesis and the set parameter for var.equal.
}
\description{
The function allows to calculate cohen's d and critical d for a two samples t-test.
}
