% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-funs.R
\name{critical_t2sp}
\alias{critical_t2sp}
\title{critical_t2sp}
\usage{
critical_t2sp(
  m1 = NULL,
  m2 = NULL,
  t = NULL,
  sd1 = NULL,
  sd2 = NULL,
  r12 = NULL,
  n,
  se = NULL,
  hypothesis = c("two.sided", "greater", "less"),
  conf.level = 0.95
)
}
\arguments{
\item{m1}{a number representing the mean of group 1.}

\item{m2}{a number representing the mean of group 2.}

\item{t}{the t value.}

\item{sd1}{a number representing the standard deviation of group 1.}

\item{sd2}{a number representing the standard deviation of group 2.}

\item{r12}{a number corresponding to the correlation between variable 1 and variable 2.}

\item{n}{a number corresponding to the sample size.}

\item{se}{a number corresponding to the standard error.}

\item{hypothesis}{a character string indicating the alternative hypothesis ("less", "greater" or "two.tailed").}

\item{conf.level}{the confidence level to set the confidence interval, default is set to 0.95.}
}
\value{
the output returns, based on the inserted values, means of the two groups, standard deviations of the two groups, numerosity, standard error, correlation between the variables, confidence level and hypothesis or t-value, numerosity, standard error, correlation between the variables, hypothesis and confidence iterval.
}
\description{
The function allows to calculate the standardized cohen's d, the critical standardized cohen's d, the cohen's d and the critical d for a paired two samples t-test.
}
