% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalvalue-methods.R
\name{critical.lm}
\alias{critical.lm}
\title{Compute critical effect size values for linear regression models}
\usage{
\method{critical}{lm}(x, conf.level = 0.95, standardize = FALSE, ...)
}
\arguments{
\item{x}{an object of class lm}

\item{conf.level}{the confidence interval level, needed to compute the smallest significant coefficient (default is 0.95, equaling a critical alpha = 0.05)}
}
\value{
an object of class critvalue
}
\description{
Compute critical effect size values for linear model coefficients of objects of class lm
}
